refactor(lowy): split Tools.hs into pure ToolCatalog + wire Tools
Two volatility axes were braided in one module: notebook-query helpers
(findNotes, getBacklinks, resolveWikilink, NoteMatch, ResolveResult) and
MCP wire scaffolding (InputSchema literals, argument parsing,
ToolHandler construction, toolJsonResult / toolError).

Phase 2 encapsulated the analogous split for resources: pure Catalog
beside MCP-typed Handlers. Tools collapsed the same seam.

Mirror the established pattern. ToolCatalog now owns the pure surface
(no dpella/mcp imports — verifiable from the import list); Tools is just
the wire adapter on top. ToolsSpec moves to ToolCatalogSpec since the
tests cover the pure helpers.
