+++ emanote/emanote.cabal
@@ -176,6 +176,7 @@ library
     Emanote.MCP.Catalog
     Emanote.MCP.Handlers
     Emanote.MCP.Server
+    Emanote.MCP.ToolCatalog
     Emanote.MCP.Tools
     Emanote.MCP.Types
     Emanote.MCP.Uri
