+++ emanote/src/Emanote/MCP/Tools.hs
@@ -1,44 +1,27 @@
-{-# LANGUAGE DuplicateRecordFields #-}
-{-# LANGUAGE NamedFieldPuns #-}
+{- | MCP wire adapter for the phase-3 query tools.
 
-{- | MCP query tools (phase 3).
+Translates between the @dpella/mcp@ 'ToolHandler' contract (JSON Schema in,
+'CallToolResult' out) and the pure queries in "Emanote.MCP.ToolCatalog".
 
-Three read-only tools exposed via MCP's @tools/list@ and @tools/call@:
+Three tools advertised:
 
-* @find_notes@ — case-insensitive substring search over titles and source paths.
-* @get_backlinks@ — wraps "Emanote.Model.Graph".'G.modelLookupBacklinks'.
-* @resolve_wikilink@ — wraps "Emanote.Model.Link.Resolve".'Resolve.resolveWikiLinkMustExist'.
+* @find_notes@ — wraps 'ToolCatalog.findNotes'.
+* @get_backlinks@ — wraps 'ToolCatalog.getBacklinks'.
+* @resolve_wikilink@ — wraps 'ToolCatalog.resolveWikilink'.
 
-Tools read from the live 'Model' snapshot via the @'IO' 'Model'@ reader
-phase 2 plumbed; no shared 'IORef', no caching.
+Tools share the live 'Model' snapshot via the @'IO' 'Model'@ reader phase 2
+plumbed; no shared 'IORef', no caching.
 -}
 module Emanote.MCP.Tools (
   tools,
-
-  -- * Pure helpers (exported for tests)
-  NoteMatch (..),
-  findNotes,
-  getBacklinks,
-  ResolveResult (..),
-  resolveWikilink,
 ) where
 
-import Commonmark.Extensions.WikiLink qualified as WL
-import Data.Aeson (ToJSON (..), (.=))
+import Data.Aeson (ToJSON, (.=))
 import Data.Aeson qualified as Aeson
-import Data.IxSet.Typed qualified as Ix
 import Data.Map.Strict qualified as Map
 import Data.Text qualified as T
-import Emanote.MCP.Uri (noteUriPrefix)
+import Emanote.MCP.ToolCatalog qualified as TC
 import Emanote.Model (Model)
-import Emanote.Model qualified as M
-import Emanote.Model.Graph qualified as G
-import Emanote.Model.Link.Rel qualified as Rel
-import Emanote.Model.Link.Resolve qualified as Resolve
-import Emanote.Model.Note qualified as N
-import Emanote.Model.StaticFile qualified as SF
-import Emanote.Model.Title qualified as Tit
-import Emanote.Route qualified as R
 import MCP.Server (
   CallToolResult,
   InputSchema (..),
@@ -48,13 +31,7 @@ import MCP.Server (
   toolTextError,
   toolTextResult,
  )
-import Network.URI.Slug qualified as Slug
-import Optics.Operators ((^.))
 import Relude
-
--- ---------------------------------------------------------------------------
--- Public entry point
--- ---------------------------------------------------------------------------
 
 -- | All MCP tools, parameterized by the model reader.
 tools :: IO Model -> [ToolHandler]
@@ -65,68 +42,8 @@ tools readModel =
   ]
 
 -- ---------------------------------------------------------------------------
--- Common types
--- ---------------------------------------------------------------------------
-
--- | A single note hit returned by 'findNotes' and 'getBacklinks'.
-data NoteMatch = NoteMatch
-  { path :: Text
-  , title :: Text
-  }
-  deriving stock (Eq, Show, Generic)
-
-{- | The @uri@ field is derived from @path@ so there is no way for the two to
-diverge: drift in 'noteUriPrefix' propagates to every consumer through
-one place.
--}
-instance ToJSON NoteMatch where
-  toJSON NoteMatch {path, title} =
-    Aeson.object
-      [ "path" .= path
-      , "title" .= title
-      , "uri" .= (noteUriPrefix <> path)
-      ]
-
-noteMatchOf :: N.Note -> NoteMatch
-noteMatchOf note =
-  NoteMatch
-    { path = toText $ R.lmlSourcePath (note ^. N.noteRoute)
-    , title = Tit.toPlain (note ^. N.noteTitle)
-    }
-
-{- | Build a 'NoteMatch' from a route. Falls back to a route-derived title
-when the note can't be looked up — used by callers that hold a route but
-not the 'N.Note' (e.g. backlink sources).
--}
-noteMatchOfRoute :: Model -> R.LMLRoute -> NoteMatch
-noteMatchOfRoute model r =
-  maybe fallback noteMatchOf (M.modelLookupNoteByRoute' r model)
-  where
-    fallback =
-      NoteMatch
-        { path = toText $ R.lmlSourcePath r
-        , title = Tit.toPlain (Tit.fromRoute r)
-        }
-
--- ---------------------------------------------------------------------------
 -- find_notes
 -- ---------------------------------------------------------------------------
-
-{- | Substring search (case-insensitive) over note titles and source paths.
-
-Returns up to @limit@ matches in 'IxSet' iteration order; this is stable
-under a given model snapshot but not lexicographically sorted. Callers that
-want ordered output should sort downstream.
--}
-findNotes :: Text -> Int -> Model -> [NoteMatch]
-findNotes query lim model =
-  let q = T.toLower query
-      hit note =
-        let m = noteMatchOf note
-         in if q `T.isInfixOf` T.toLower (title m) || q `T.isInfixOf` T.toLower (path m)
-              then Just m
-              else Nothing
-   in take (max 0 lim) $ mapMaybe hit $ Ix.toList (model ^. M.modelNotes)
 
 findNotesTool :: IO Model -> ToolHandler
 findNotesTool readModel =
@@ -149,24 +66,11 @@ findNotesTool readModel =
         Right q -> do
           let lim = fromMaybe 20 $ readIntArg "limit" margs
           model <- liftIO readModel
-          pure $ toolJsonResult (Aeson.object ["matches" .= findNotes q lim model])
+          pure $ toolJsonResult (Aeson.object ["matches" .= TC.findNotes q lim model])
 
 -- ---------------------------------------------------------------------------
 -- get_backlinks
 -- ---------------------------------------------------------------------------
-
-{- | Backlinks for the note at the given source path.
-
-Returns 'Left' if @path@ isn't a recognised LML source path
-(@guide/mcp.md@, @daily/2024-01-01.org@, …). An empty list is a valid
-'Right' result and means the note exists but no other note links to it.
--}
-getBacklinks :: FilePath -> Model -> Either Text [NoteMatch]
-getBacklinks fp model =
-  case R.mkLMLRouteFromMdOrOrgFilePath fp of
-    Nothing -> Left $ "Not a recognised note path: " <> toText fp
-    Just r ->
-      Right $ noteMatchOfRoute model . fst <$> G.modelLookupBacklinks r model
 
 getBacklinksTool :: IO Model -> ToolHandler
 getBacklinksTool readModel =
@@ -187,60 +91,13 @@ getBacklinksTool readModel =
         Left err -> pure $ toolError err
         Right p -> do
           model <- liftIO readModel
-          pure $ case getBacklinks (toString p) model of
+          pure $ case TC.getBacklinks (toString p) model of
             Left err -> toolError err
             Right ms -> toolJsonResult (Aeson.object ["backlinks" .= ms])
 
 -- ---------------------------------------------------------------------------
 -- resolve_wikilink
 -- ---------------------------------------------------------------------------
-
--- | Outcome of resolving a wikilink, mirroring 'Rel.ResolvedRelTarget'.
-data ResolveResult
-  = ResolvedNote NoteMatch
-  | ResolvedStatic Text
-  | UnresolvedMissing
-  | UnresolvedAmbiguous [Either NoteMatch Text]
-  deriving stock (Eq, Show, Generic)
-
-instance ToJSON ResolveResult where
-  toJSON = \case
-    ResolvedNote nm ->
-      Aeson.object ["result" .= ("found" :: Text), "kind" .= ("note" :: Text), "note" .= nm]
-    ResolvedStatic p ->
-      Aeson.object ["result" .= ("found" :: Text), "kind" .= ("static" :: Text), "path" .= p]
-    UnresolvedMissing ->
-      Aeson.object ["result" .= ("missing" :: Text)]
-    UnresolvedAmbiguous cs ->
-      Aeson.object
-        [ "result" .= ("ambiguous" :: Text)
-        , "candidates" .= (candidateValue <$> cs)
-        ]
-    where
-      candidateValue = \case
-        Left nm -> Aeson.object ["kind" .= ("note" :: Text), "note" .= nm]
-        Right p -> Aeson.object ["kind" .= ("static" :: Text), "path" .= p]
-
-{- | Resolve a wikilink string (without brackets), optionally relative to a
-source note for ambiguity disambiguation. Defaults the @from@ context to
-the notebook index when unspecified.
--}
-resolveWikilink :: Text -> Maybe FilePath -> Model -> Either Text ResolveResult
-resolveWikilink wlText mFromPath model = do
-  wl <- maybeToRight ("Not a valid wikilink: " <> wlText) (parseWikiLinkText wlText)
-  fromR <- case mFromPath of
-    Nothing -> Right (M.modelIndexRoute model)
-    Just p -> maybeToRight ("Not a recognised note path: " <> toText p) (R.mkLMLRouteFromMdOrOrgFilePath p)
-  Right $ case Resolve.resolveWikiLinkMustExist model fromR wl of
-    Rel.RRTFound (Left (_, note)) -> ResolvedNote (noteMatchOf note)
-    Rel.RRTFound (Right sf) -> ResolvedStatic (staticFilePath sf)
-    Rel.RRTMissing -> UnresolvedMissing
-    Rel.RRTAmbiguous cs -> UnresolvedAmbiguous $ toList $ candidate <$> cs
-  where
-    candidate = \case
-      Left (_, note) -> Left (noteMatchOf note)
-      Right sf -> Right (staticFilePath sf)
-    staticFilePath sf = toText $ R.encodeRoute (sf ^. SF.staticFileRoute)
 
 resolveWikilinkTool :: IO Model -> ToolHandler
 resolveWikilinkTool readModel =
@@ -263,19 +120,13 @@ resolveWikilinkTool readModel =
         Right wl -> do
           let mFrom = toString <$> readTextArgMaybe "from" margs
           model <- liftIO readModel
-          pure $ case resolveWikilink wl mFrom model of
+          pure $ case TC.resolveWikilink wl mFrom model of
             Left err -> toolError err
             Right res -> toolJsonResult res
 
 -- ---------------------------------------------------------------------------
--- Helpers
+-- Wire helpers
 -- ---------------------------------------------------------------------------
-
--- | Parse a slash-separated wikilink target (e.g. "foo/bar") into a 'WL.WikiLink'.
-parseWikiLinkText :: Text -> Maybe WL.WikiLink
-parseWikiLinkText s
-  | T.null s = Nothing
-  | otherwise = viaNonEmpty WL.mkWikiLinkFromSlugs (Slug.decodeSlug <$> T.splitOn "/" s)
 
 toolJsonResult :: (ToJSON a) => a -> ProcessResult CallToolResult
 toolJsonResult v = ProcessSuccess $ toolTextResult [decodeUtf8 (Aeson.encode v)]
